Refaktor Registerkontrolltreff-komponent slik at den blir enklere å gjenbruke for Utpeking av Norge
