--- src/sider/registrering/komponenter/registerkontrollTreff.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import PT from 'prop-types';
-import MKV from '../../../melosyskodeverk';
-
-import * as KV from '../../../kodeverk';
-import * as Nav from '../../../utils/navFrontend';
-
-const UnntakPeriodeBegrunnelse = kode => {
-  if (!kode) return '';
-  return KV.kodeTilTerm(kode, MKV.KTObjects.begrunnelser.kontroll_begrunnelser);
-};
-
-const RegisterkontrollTreff = ({ begrunnelseKode }) => (
-  <div className="registerkontroll-listeelement">
-    <Nav.Ikoner kind="advarsel-sirkel-fyll" size="24" />
-    <Nav.typo.Normaltekst>{UnntakPeriodeBegrunnelse(begrunnelseKode)}</Nav.typo.Normaltekst>
-  </div>
-);
-
-RegisterkontrollTreff.propTypes = {
-  begrunnelseKode: PT.string.isRequired,
-};
-
-export default RegisterkontrollTreff;
